double file was deleted
